+++ src/hittable/triangle.rs
@@ -31,12 +31,12 @@ impl<'a> Triangle<'a> {
 }
 
 impl<'a> Hittable for Triangle<'a> {
-    fn hit(&self, ray: &Ray, t_min: f32, t_max: f32, _: &mut FastRng) -> Hit {
+    fn hit(&self, ray: &Ray, t_min: f64, t_max: f64, _: &mut FastRng) -> Hit {
         let u = ray.direction;
         let (v, w) = self.edges;
         let normal = v.cross(w);
         let determinant = normal.dot(u);
-        if determinant.abs() < 1e-8_f32 {
+        if determinant.abs() < 1e-8_f64 {
             None
         } else {
             let a_minus_origin = self.vertex - ray.origin;
@@ -67,8 +67,8 @@ impl<'a> Hittable for Triangle<'a> {
     fn bounding_box(&self) -> BoundingBox {
         let vertices = self.vertices();
         BoundingBox {
-            minimum: vertices.0.min(vertices.1).min(vertices.2) - 1e-7f32,
-            maximum: vertices.0.max(vertices.1).max(vertices.2) + 1e-7f32,
+            minimum: vertices.0.min(vertices.1).min(vertices.2) - 1e-7f64,
+            maximum: vertices.0.max(vertices.1).max(vertices.2) + 1e-7f64,
         }
     }
 }
